+++ app/controllers/sports_controller.rb
@@ -4,4 +4,11 @@ class SportsController < BaseController
 
   MODEL = Sport
 
+  def view(index)
+    sport = get_by_index(index.to_i)
+    if sport
+      
+    end
+  end
+
 end