+++ app/controllers/teams_controller.rb
@@ -1,0 +1,11 @@
+require_relative "base_controller"
+
+class TeamsController < BaseController
+
+  MODEL = Team
+
+  def initialize(sport)
+    @parent = sport
+  end
+
+end