+++ db/migrate/201408111339_create_teams.rb
@@ -1,0 +1,9 @@
+class CreateTeams < ActiveRecord::Migration
+
+  def change
+    create_table "teams", force: true do |t|
+      t.string "name"
+    end
+  end
+
+end