+++ db/migrate1/201408111702_add_association_columns.rb
@@ -1,0 +1,10 @@
+class AddAssociationColumns < ActiveRecord::Migration
+
+  def change
+    # add_column :sports, :teams, :has_many
+    # add_column :teams, :sport, :belongs_to
+    # add_column :teams, :players, :has_many
+    # add_column :teams, :captain, :has_one
+  end
+
+end