+++ spec/integration/viewing_sports.rb
@@ -1,0 +1,15 @@
+RSpec.describe "Viewing sports", :integration do
+  context "with valid input" do
+    let!(:sport){Sport.create(name: "Baseball")}
+    let!(:output){run_program_with_input("1")}
+
+    it "prints the sport into the terminal" do
+      header = <<EOS
+==============
+Baseball
+==============
+EOS
+      expect(output).to include(header)
+    end
+  end
+end